# swagger.yaml
basePath: /
consumes:
- text/plain
definitions:
...
  addWorldTendencyReq:
    properties:
      Area1:
        format: int64
        type: integer
      Area2:
        format: int64
        type: integer
      Area3:
        format: int64
        type: integer
      Area4:
        format: int64
        type: integer
      Area5:
        format: int64
        type: integer
      Area6:
        format: int64
        type: integer
      Area7:
        format: int64
        type: integer
      CharacterID:
        type: string
      LR1:
        format: int64
        type: integer
      LR2:
        format: int64
        type: integer
      LR3:
        format: int64
        type: integer
      LR4:
        format: int64
        type: integer
      LR5:
        format: int64
        type: integer
      LR6:
        format: int64
        type: integer
      LR7:
        format: int64
        type: integer
      WB1:
        format: int64
        type: integer
      WB2:
        format: int64
        type: integer
      WB3:
        format: int64
        type: integer
      WB4:
        format: int64
        type: integer
      WB5:
        format: int64
        type: integer
      WB6:
        format: int64
        type: integer
      WB7:
        format: int64
        type: integer
    type: object
    x-go-package: github.com/danmrichards/dessego/internal/server/game
  bloodMsgGradeReq:
    properties:
      CharacterID:
        type: string
      Version:
        format: int64
        type: integer
    type: object
    x-go-package: github.com/danmrichards/dessego/internal/server/game
...
  getGhostReq:
    properties:
      BlockID:
        format: uint32
        type: integer
      CharacterID:
        type: string
      MaxGhosts:
        format: int64
        type: integer
      SOS:
        format: int64
        type: integer
      SOSIDList:
        items:
          type: string
        type: array
      Version:
        format: int64
        type: integer
    type: object
    x-go-package: github.com/danmrichards/dessego/internal/server/game
...
  initCharacterReq:
    properties:
      CharacterID:
        type: string
      Index:
        format: int64
        type: integer
      Version:
        format: int64
        type: integer
    type: object
    x-go-package: github.com/danmrichards/dessego/internal/server/game
...
  multiplayerGradeReq:
    properties:
      CharacterID:
        type: string
      Version:
        format: int64
        type: integer
    type: object
    x-go-package: github.com/danmrichards/dessego/internal/server/game
...
  setGhostReq:
    properties:
      CharacterID:
        type: string
      GhostBlockID:
        format: uint32
        type: integer
      PosX:
        format: float
        type: number
      PosY:
        format: float
        type: number
      PosZ:
        format: float
        type: number
      ReplayData:
        type: string
    type: object
    x-go-package: github.com/danmrichards/dessego/internal/server/game
...
  worldTendencyReq:
    properties:
      MaxNum:
        format: int64
        type: integer
      Version:
        format: int64
        type: integer
    type: object
    x-go-package: github.com/danmrichards/dessego/internal/server/game
info:
  description: The boostrap and game API endpoints for Demon's Souls
  title: Demon's Souls API
  version: 1.0.0
paths:
...
  /cgi-bin/addWorldTendencyHandler.spd:
    post:
      consumes:
      - text/plain
      description: Updates the currently stored world tendency data for a given character
      operationId: addWorldTendencyHandler
      parameters:
      - in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/addWorldTendencyReq'
      produces:
      - text/plain
      responses:
        "200":
          description: successful operation
        "500":
          description: unsuccessful operation
      summary: Add world tendency
      tags:
      - character
...
  /cgi-bin/getBloodMessageGrade.spd:
    post:
      consumes:
      - text/plain
      description: Gets the current blood message grade for a character
      operationId: getBloodMessageGrade
      parameters:
      - in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/bloodMsgGradeReq'
      produces:
      - text/plain
      responses:
        "200":
          description: successful operation
        "500":
          description: unsuccessful operation
      summary: Get blood message grade
      tags:
      - character
  /cgi-bin/getMultiPlayGrade.spd:
    post:
      consumes:
      - text/plain
      description: Gets the current multiplayer grade for a character, used when matchmaking
      operationId: getMultiPlayGrade
      parameters:
      - in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/multiplayerGradeReq'
      produces:
      - text/plain
      responses:
        "200":
          description: successful operation
        "500":
          description: unsuccessful operation
      summary: Get multiplayer grade
      tags:
      - character
  /cgi-bin/getQWCData.spd:
    post:
      consumes:
      - text/plain
      description: Returns the currently stored world tendency data for a given character
      operationId: getQWCData
      parameters:
      - in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/worldTendencyReq'
      produces:
      - text/plain
      responses:
        "200":
          description: successful operation
        "500":
          description: unsuccessful operation
      summary: Get world tendency
      tags:
      - character
...
  /cgi-bin/getWanderingGhost.spd:
    post:
      consumes:
      - text/plain
      description: Returns a list of wandering ghosts (replays) within a given area
        of the game
      operationId: getWanderingGhost
      parameters:
      - in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/getGhostReq'
      produces:
      - text/plain
      responses:
        "200":
          description: successful operation
        "500":
          description: unsuccessful operation
      summary: Get wandering ghost
      tags:
      - ghost
  /cgi-bin/initializeCharacter.spd:
    post:
      consumes:
      - text/plain
      description: Initialises a new character and persists it for future reference
      operationId: initializeCharacter
      parameters:
      - in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/initCharacterReq'
      produces:
      - text/plain
      responses:
        "200":
          description: successful operation
        "500":
          description: unsuccessful operation
      summary: Initialises a new character
      tags:
      - character
...
  /cgi-bin/setWanderingGhost.spd:
    post:
      consumes:
      - text/plain
      description: Stores wandering ghost (replay) data for the current player
      operationId: setWanderingGhost
      parameters:
      - in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/setGhostReq'
      produces:
      - text/plain
      responses:
        "200":
          description: successful operation
        "500":
          description: unsuccessful operation
      summary: Set wandering ghost
      tags:
      - ghost
...
produces:
- text/plain
schemes:
- http
swagger: "2.0"